Add upload provider and bucket name as required helm values (#21)
The provider and bucket name are now required configuration values for
edge sites. Rather than require the user to configure this information
in the UI we require this configuration in the values when the
deployment is setup. The user must specify this information when
deploying their edge site.

Related data platform change:
https://github.com/foxglove/data-platform/pull/808/