Add deployment template for site controller
Adds a deployment template for the site controller, with a metrics port
associated with the pod. Enables users to put custom pod labels on the
inbox listener, stream service, and site controller, so that they may
opt these deployments into scraping executed by some external framework
on the basis of labels.

For instance, if a prometheus PodMonitor is set up to scrape pods with
label monitoring: true, this can now be accomplished through the chart.