Only source GOOGLE_APPLICATION_CREDENTIALS if secret exists
This makes the deployment runnable in minikube with passthrough GCP
credentials. Prior to this commit, when the secret doesn't exist, the
app will source an empty value for GOOGLE_APPLICATION_CREDENTIALS, which
will kick the GCP talking components into the env credential provider
and defeat providers (CLI, machine creds) you may be attempting to use.