Edge controller: only define GCP credentials env var when secret exists
This is what we do for the other charts as well, and makes it possible
to use the edge controller from e.g minikube using CLI credentials in a
testing context. This is convenient for development and for potential
users who just want to kick the tires without provisioning service
accounts.