FG-5079: Update Stream Service Probes (#67)
### Public-Facing Changes

Updates Stream Service probes to allow for a longer window before
terminating the pod.

### Description

The default failureThreshold for Kubernetes probes is 3, meaning the
stream service would be terminated after ~12 seconds of not responding
to requests. This change ups that threshold to 10 failures before the
pod is removed from taking requests and 20 failures before it is
terminated.