Update primary site to default to beta stream service (#112)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

The chart now defaults to the new stream server image. There is a flag
to opt out, however this will eventually be removed.

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

Docs PR incoming.

### Description

This change updates the helm charts to default to the new version of the
stream server. It also adds in an option that can be enabled to revert
back to the old stream server.