primary-site: update to v0.0.67 (#123)
### Changelog
- `inbox-listener`: added `max_finalization_worker_count` environment
variable to configure the maximum worker count used to write files to
the lake. This defaults to 40 and can be reduced to mitigate
rate-limiting API errors from the lake bucket. When set to 0, the inbox
listener uses one worker per topic in the input file.
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

### Docs

https://github.com/foxglove/docs/pull/678
<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->