primary-site: update to version 0.0.71 (#128)
### Changelog
None.

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description
- Includes a template change to allow toggling to the new Rust inbox
listener image. This flag is opt-in for this version. In a later version
it will be switched to opt-out.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

### Manual testing
I checked that the templating produces the legacy inbox listener image
with `helm template charts/primary-site`.


<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->