Add simple tests for primary site chart (#137)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Adds in some simple tests that will run on pull requests and validate
that the helm chart installs correctly and the pods are created.

Fixes FIRE-6.