primary-site: bump to v0.0.86 (#150)
### Changelog
- `query-server`: add success JSON to `/` endpoint

### Docs

None.

### Description

Adds a success JSON document to the `/` response body from the query
service, which will go with instructions in installation docs to use
this endpoint to check that the ingress works.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->